+++ src/renderer/dto/render-command.dto.ts
@@ -1,7 +1,10 @@
+import { IsUrl } from "class-validator"
 import { HttpProxy } from "../../common/types/http-proxy.class"
 
 export class RenderCommandDto {
+  @IsUrl()
   url: string
+
   blockAds?: boolean = true
   headers?: Record<string, string>
   httpProxy?: HttpProxy