+++ src/renderer/dto/render-query.dto.ts
@@ -1,0 +1,4 @@
+import { PickType } from "@nestjs/swagger"
+import { RenderCommandDto } from "./render-command.dto"
+
+export class RenderQueryDto extends PickType(RenderCommandDto, ["url", "blockAds"]) {}